requirements.txt
pdf2image==1.16.3
Pillow==10.0.0
Flask==2.3.3
Werkzeug==2.3.7
